add the sam tutotrial
