+++ multimodal/src/autogluon/multimodal/utils/__init__.py
@@ -65,7 +65,7 @@ from .metric import (
     infer_metrics,
     infer_problem_type_by_eval_metric,
 )
-from .misc import logits_to_prob, merge_bio_format, shopee_dataset, tensor_to_ndarray, visualize_ner
+from .misc import logits_to_prob, merge_bio_format, shopee_dataset, tensor_to_ndarray
 from .mmcv import CollateMMDet, CollateMMOcr
 from .model import (
     create_fusion_model,
@@ -93,6 +93,6 @@ from .object_detection import (
     save_result_voc_format,
     visualize_detection,
 )
-from .object_detection_visualizer import Visualizer
 from .ovd import save_ovd_result_df
 from .save import process_save_path, save_text_tokenizers, setup_save_path
+from .visualizer import NERVisualizer, ObjectDetectionVisualizer, SemanticSegmentationVisualizer, visualize_ner
