+++ multimodal/src/autogluon/multimodal/utils/misc.py
@@ -1,4 +1,3 @@
-import collections
 import json
 import logging
 import os
@@ -151,90 +150,3 @@ def merge_bio_format(data, preds):
             ]
         )
     return results
-
-
-class NERVisualizer:
-    """An NER visualizer that renders NER prediction as a string of HTML
-    inline to any Python class Jupyter notebooks.
-    """
-
-    def __init__(self, pred, sent, seed):
-        self.pred = pred
-        self.sent = sent
-        self.colors = {}
-        self.spans = merge_spans(sent, pred, for_visualizer=True)
-        self.spans = collections.OrderedDict(sorted(self.spans.items()))
-        self.rng = np.random.RandomState(seed)
-
-    @staticmethod
-    def escape_html(text: str) -> str:
-        """Replace <, >, &, " with their HTML encoded representation. Intended to
-        prevent HTML errors in rendered displaCy markup.
-        text (str): The original text.
-        RETURNS (str): Equivalent text to be safely used within HTML.
-        """
-        text = text.replace("&", "&amp;")
-        text = text.replace("<", "&lt;")
-        text = text.replace(">", "&gt;")
-        text = text.replace('"', "&quot;")
-        return text
-
-    def html_template(self, text, label, color):
-        """
-        Generate an HTML template for the given text and its label.
-
-        Parameters
-        ----------
-        text
-            The text to be highlighted.
-        label
-            The predicted label for the given text.
-        color
-            The background color of the mark tag.
-        """
-        text = '<mark style="background-color:{}; color:white; border-radius: .6em .6em; padding: .1em;">{} \
-         <b style="background-color:white; color:black; font-size:x-small; border-radius: 0.5em .5em; padding: .0em;">{} </b> \
-         </mark>'.format(
-            color, self.escape_html(text), self.escape_html(label)
-        )
-        return text
-
-    def _repr_html_(self):
-        entities = []
-        new_sent = ""
-        last = 0
-        for key, value in self.spans.items():
-            entity_group = value[-1]
-            if re.match("B-", entity_group, re.IGNORECASE) or re.match("I-", entity_group, re.IGNORECASE):
-                entity_group = entity_group[2:]
-            if entity_group not in self.colors:
-                self.colors.update({entity_group: "#%06X" % self.rng.randint(0, 0xFFFFFF)})
-            start = key
-            new_sent += self.sent[last:start]
-            last = end = value[0]
-            entity_text = self.html_template(self.sent[start:end], entity_group, color=self.colors[entity_group])
-            new_sent += entity_text
-        new_sent += self.sent[last:]
-
-        return new_sent
-
-
-def visualize_ner(sentence, prediction, seed=0):
-    """
-    Visualize the prediction of NER.
-
-    Parameters
-    ----------
-    sentence
-        The input sentence.
-    prediction
-        The NER prediction for the sentence.
-    seed
-        The seed for colorpicker.
-
-    Returns
-    -------
-    An NER html visualizer.
-    """
-    visualizer = NERVisualizer(prediction, sentence, seed)
-    return visualizer
